app/requirements.txt
huggingface_hub==0.22.2
gradio==3.35.2
gradio_client==0.2.9
numpy
torch
transformers==4.31.0
pydub
faster_whisper
pyannote.audio